unskip spectral extraction test, move test from catalogs plugin to catalogs loader (#4329)
